+++ services/requester/cross-spork_client_test.go
@@ -1,22 +1,28 @@
-package models
+package requester
 
 import (
 	"context"
+	"testing"
+
+	"github.com/onflow/flow-go-sdk/access"
+	"github.com/onflow/flow-go-sdk/access/grpc"
 	"github.com/rs/zerolog"
 	"github.com/stretchr/testify/require"
-	"testing"
 )
 
 func TestCrossSporkClient_MultiClient(t *testing.T) {
+	sporkClients := []access.Client{
+		grpc.NewClient("test1.com"),
+	}
 	clientHosts := []string{"test1.com", "test2.com", "test3.com"}
 
 	client, err := NewCrossSporkClient(clientHosts[0], zerolog.Nop())
 	require.NoError(t, err)
 
-	err = client.AddSpork(100, clientHosts[1])
+	err = client.addSpork(clientHosts[1])
 	require.NoError(t, err)
 
-	err = client.AddSpork(200, clientHosts[2])
+	err = client.addSpork(clientHosts[2])
 	require.NoError(t, err)
 
 	c := client.getClientForHeight(300)
@@ -45,9 +51,9 @@ func TestCrossSporkClient_ExistingHeight(t *testing.T) {
 	client, err := NewCrossSporkClient("host1.com", zerolog.Nop())
 	require.NoError(t, err)
 
-	err = client.AddSpork(100, "host2.com")
+	err = client.addSpork(100, "host2.com")
 	require.NoError(t, err)
 
-	err = client.AddSpork(100, "host3.com")
+	err = client.addSpork(100, "host3.com")
 	require.EqualError(t, err, "provided last height already exists")
 }
